Add package entry point
