+++ zrad/__main__.py
@@ -1,0 +1,6 @@
+"""Run the Z-Rad GUI as a module."""
+
+from .run_gui import main
+
+if __name__ == "__main__":
+    main()
